Create Transform module
Contains: Operation, Window, WOperation, BinProperty.
