+++ hvega/hvega.cabal
@@ -132,6 +132,7 @@ library
                        Graphics.Vega.VegaLite.Selection
                        Graphics.Vega.VegaLite.Specification
                        Graphics.Vega.VegaLite.Time
+                       Graphics.Vega.VegaLite.Transform
   build-depends:       base >= 4.9 && < 5
                      , aeson >= 0.11 && < 1.5
                      , text == 1.2.*
