+++ hvega/src/Graphics/Vega/VegaLite.hs
@@ -613,6 +613,7 @@ import qualified Graphics.Vega.VegaLite.Output as VL
 import qualified Graphics.Vega.VegaLite.Selection as VL
 import qualified Graphics.Vega.VegaLite.Specification as VL
 import qualified Graphics.Vega.VegaLite.Time as VL
+import qualified Graphics.Vega.VegaLite.Transform as VL
 
 
 -- Documentation
